refactor(store-core): extract dev-preview handlers into handlers/dev-preview.ts (audit P2-3)
Move the dev-preview family (handleDevPreview, handleDevPreviewStopped +
DevPreviewBuilder) out of the handlers/index.ts god-module into
handlers/dev-preview.ts, behind the barrel's export * re-export so the public
surface is unchanged.

Drop the now-unused DevPreview type import from index.ts (it was only used by
the moved handlers).

Pure move — body byte-identical to origin/main. index.ts 1992 -> 1931 lines.
