+++ packages/store-core/src/handlers/index.ts
@@ -12,7 +12,6 @@
 import type {
   ActiveTool,
   ChatMessage,
-  DevPreview,
   ModelInfo,
   TranscriptBackgroundTask,
   ServerError,
@@ -373,73 +372,11 @@ export function applyInterventionBuilder(
 }
 
 // ---------------------------------------------------------------------------
-// dev_preview / dev_preview_stopped
-//
-// These handlers are stateful in a way the others aren't: the new devPreviews
-// array depends on the existing array (dedup-by-port for `dev_preview`, filter
-// for `dev_preview_stopped`). To keep a single sessionId resolution path
-// (matching plan_started/plan_ready), the handlers return a builder shape:
-// `sessionId` resolved as usual, plus an `applyTo(current)` function the call
-// site invokes with the looked-up array. This avoids the double-resolution
-// pattern that would otherwise be needed if `currentPreviews` were a
-// pre-handler argument.
+// Dev-preview handlers (dev_preview, dev_preview_stopped) live in
+// ./dev-preview.ts (audit P2-3 split). Re-exported here so the barrel's public
+// surface is unchanged.
 // ---------------------------------------------------------------------------
-
-/** Builder result for handlers whose patch depends on existing state. */
-export interface DevPreviewBuilder {
-  /** Session ID the patch targets (may be null if no session context). */
-  sessionId: string | null
-  /** Apply the builder to the session's current devPreviews. */
-  applyTo: (current: DevPreview[]) => { devPreviews: DevPreview[] }
-}
-
-/**
- * Resolve target session and produce a builder that appends (or replaces by
- * port) a dev-preview entry. Both clients dedupe by port: a new preview for
- * an already-tracked port replaces the existing entry, otherwise it is
- * appended after the filtered remainder.
- *
- * Behaviour-preserving: `msg.port` and `msg.url` are forwarded verbatim with
- * the same unsafe cast (`port as number, url as string`) the prior inline
- * implementations used. Tightening to runtime validation would be a behaviour
- * change and is out of scope for the #2661 mechanical migration.
- */
-export function handleDevPreview(
-  msg: Record<string, unknown>,
-  activeSessionId: string | null,
-): DevPreviewBuilder {
-  const preview: DevPreview = {
-    port: msg.port as number,
-    url: msg.url as string,
-  }
-  return {
-    sessionId: resolveSessionId(msg, activeSessionId),
-    applyTo: (current) => ({
-      devPreviews: [...current.filter((p) => p.port !== preview.port), preview],
-    }),
-  }
-}
-
-/**
- * Resolve target session and produce a builder that removes the dev-preview
- * entry matching `msg.port`. If no entry matches the previews list is returned
- * unchanged (matches both clients' prior `filter`-based inline implementation).
- *
- * Behaviour-preserving: `msg.port` is cast verbatim (`port as number`) without
- * runtime validation, matching the prior inline behaviour.
- */
-export function handleDevPreviewStopped(
-  msg: Record<string, unknown>,
-  activeSessionId: string | null,
-): DevPreviewBuilder {
-  const stoppedPort = msg.port as number
-  return {
-    sessionId: resolveSessionId(msg, activeSessionId),
-    applyTo: (current) => ({
-      devPreviews: current.filter((p) => p.port !== stoppedPort),
-    }),
-  }
-}
+export * from './dev-preview'
 
 // ---------------------------------------------------------------------------
 // auth_ok / auth_fail / key_exchange_ok / server_mode
